move post_xform to test_utils
