+++ corehq/ex-submodules/couchforms/tests/test_auth.py
@@ -4,8 +4,7 @@ from corehq.util.test_utils import TestFileMixin
 from couchforms.models import DefaultAuthContext
 import os
 
-from corehq.form_processor.test_utils import run_with_all_backends
-from corehq.form_processor.interfaces.processor import FormProcessorInterface
+from corehq.form_processor.test_utils import run_with_all_backends, post_xform
 
 
 class AuthTest(TestCase, TestFileMixin):
@@ -19,5 +18,5 @@ class AuthTest(TestCase, TestFileMixin):
         def process(xform):
             xform.auth_context = DefaultAuthContext().to_json()
 
-        xform = FormProcessorInterface().post_xform(xml_data, process=process)
+        xform = post_xform(xml_data, process=process)
         self.assertEqual(xform.auth_context, {'doc_type': 'DefaultAuthContext'})
