+++ corehq/ex-submodules/couchforms/tests/test_duplicates.py
@@ -3,8 +3,7 @@ from django.test import TestCase
 
 from corehq.apps.receiverwrapper import submit_form_locally
 from corehq.form_processor.interfaces.dbaccessors import FormAccessors
-from corehq.form_processor.interfaces.processor import FormProcessorInterface
-from corehq.form_processor.test_utils import FormProcessorTestUtils, run_with_all_backends
+from corehq.form_processor.test_utils import FormProcessorTestUtils, run_with_all_backends, post_xform
 from corehq.util.test_utils import TestFileMixin
 
 
@@ -13,21 +12,18 @@ class DuplicateFormTest(TestCase, TestFileMixin):
     file_path = ('data', 'posts')
     root = os.path.dirname(__file__)
 
-    def setUp(self):
-        self.interface = FormProcessorInterface()
-
     def tearDown(self):
         FormProcessorTestUtils.delete_all_xforms()
 
     @run_with_all_backends
     def test_basic_duplicate(self):
         xml_data = self.get_xml('duplicate')
-        xform = self.interface.post_xform(xml_data)
+        xform = post_xform(xml_data)
         self.assertEqual(self.ID, xform.form_id)
         self.assertTrue(xform.is_normal)
         self.assertEqual("test-domain", xform.domain)
 
-        xform = self.interface.post_xform(xml_data, domain='test-domain')
+        xform = post_xform(xml_data, domain='test-domain')
         self.assertNotEqual(self.ID, xform.form_id)
         self.assertTrue(xform.is_duplicate)
         self.assertTrue(self.ID in xform.problem)
@@ -38,7 +34,7 @@ class DuplicateFormTest(TestCase, TestFileMixin):
         instance = self.get_xml('duplicate')
 
         # Post an xform with an alternate doc_type
-        xform1 = self.interface.post_xform(
+        xform1 = post_xform(
             instance,
             domain=domain,
         )
