+++ corehq/ex-submodules/couchforms/tests/test_edits.py
@@ -15,7 +15,7 @@ from couchforms.models import (
 )
 
 from corehq.form_processor.interfaces.processor import FormProcessorInterface
-from corehq.form_processor.test_utils import FormProcessorTestUtils, run_with_all_backends
+from corehq.form_processor.test_utils import FormProcessorTestUtils, run_with_all_backends, post_xform
 from corehq.util.test_utils import TestFileMixin
 
 
@@ -46,13 +46,13 @@ class EditFormTest(TestCase, TestFileMixin):
             form.domain = self.domain
             form.received_on = yesterday
 
-        xform = self.interface.post_xform(original_xml, process=process)
+        xform = post_xform(original_xml, process=process)
         self.assertEqual(self.ID, xform.form_id)
         self.assertTrue(xform.is_normal)
         self.assertEqual("", xform.form_data['vitals']['height'])
         self.assertEqual("other", xform.form_data['assessment']['categories'])
 
-        xform = self.interface.post_xform(edit_xml, domain=self.domain)
+        xform = post_xform(edit_xml, domain=self.domain)
         self.assertEqual(self.ID, xform.form_id)
         self.assertTrue(xform.is_normal)
         self.assertEqual("100", xform.form_data['vitals']['height'])
