+++ corehq/form_processor/backends/couch/processor.py
@@ -9,7 +9,6 @@ from casexml.apps.case.models import CommCareCase, CommCareCaseAction
 from casexml.apps.case.util import get_case_xform_ids
 from casexml.apps.case.xform import get_case_updates
 from corehq.form_processor.exceptions import CaseNotFound
-from corehq.form_processor.parsers.form import process_xform
 from couchforms.util import fetch_and_wrap_form
 from couchforms.models import (
     XFormInstance, XFormDeprecated, XFormDuplicate,
@@ -20,23 +19,6 @@ from corehq.form_processor.utils import extract_meta_instance_id
 
 
 class FormProcessorCouch(object):
-
-    @classmethod
-    def post_xform(cls, instance_xml, attachments=None, process=None, domain='test-domain'):
-        """
-        create a new xform and releases the lock
-
-        this is a testing entry point only and is not to be used in real code
-
-        """
-        if not process:
-            def process(xform):
-                xform.domain = domain
-        xform_lock = process_xform(domain, instance_xml, attachments=attachments, process=process)
-        with xform_lock as xforms:
-            for xform in xforms:
-                xform.save()
-            return xforms[0]
 
     @classmethod
     def store_attachments(cls, xform, attachments):
